% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_LD.R
\name{make_LD}
\alias{make_LD}
\title{Create a new class for longitudinal data}
\usage{
make_LD(dat)
}
\arguments{
\item{dat}{a data frame which collected over times on a given subject/person at multiple visits and locations.}
}
\value{
returns a list with class \code{"LongitudinalData"} containing the following variables: id, visit, room, value and timepoint.
}
\description{
Create a new class for longitudinal data
}
\examples{
library(readr)
mie <- read_csv("D:/2018_1st_term/stats_computing/hw/hw3/MIE.csv.gz", col_types = "iicdi")
x = make_LD(mie)
class(x)
print(x)

}
