% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room.R
\name{room}
\alias{room}
\title{Extract room-specific information}
\usage{
room(x, room_name)
}
\arguments{
\item{x}{an object of class \code{"LongitudinalData"}}

\item{room_name}{the specific room that you select}
}
\value{
an object of class \code{"room"}
}
\description{
Extract room-specific information
}
\examples{
\dontrun{

out_idvr= subject(x, 44) \%>\% visit(0) \%>\% room("bedroom")
print(out_idvr)

out_sum1 <- subject(x, 44) \%>\% visit(0) \%>\% room("bedroom") \%>\% summary
print(out_sum1)

out_summary <- subject(x, 44) \%>\% visit(1) \%>\% room("living room") \%>\% summary
print(out_summary)
}
}
