% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{subject}
\alias{subject}
\title{Extract subject-specific information}
\usage{
subject(x, id_num)
}
\arguments{
\item{x}{an object of class \code{"LongitudinalData"}}

\item{id_num}{the specific subject id that you select}
}
\value{
an object of class \code{"subject"}
}
\description{
Extract subject-specific information
}
\examples{
\dontrun{

out = subject(x,54)
print(out)

out_summary= subject(x,54) \%>\% summary
print(out_summary)
}
}
