% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{visit}
\alias{visit}
\title{Extract visit-specific information}
\usage{
visit(x, visit_num)
}
\arguments{
\item{x}{an object of class \code{"LongitudinalData"}}

\item{visit_num}{the specific visit that you select}
}
\value{
an object of class \code{"visit"}
}
\description{
Extract visit-specific information
}
\examples{

\dontrun{
out_idv = subject(x,44)\%>\%visit(0)
print(out_idv)
}



}
